* Add environment variables to docker-compose.yml * Remover environment variables from Dockerfile * Unit test moved to UseCase folder * SingInUser unit test improvments
