+++ Metafar.Challenge/Dockerfile
@@ -2,8 +2,6 @@
 USER $APP_UID
 WORKDIR /app
 
-#ENV ASPNETCORE_URLS=http://+:5000
-
 FROM mcr.microsoft.com/dotnet/sdk:8.0 AS build
 ARG BUILD_CONFIGURATION=Release
 WORKDIR /src
