+++ Metafar.Challenge/Metafar.Challenge.UnitTest/UseCases/SignInUserByCardUnitTest.cs
@@ -1,18 +1,14 @@
 using Metafar.Challenge.Dto;
 using Metafar.Challenge.Entity;
 using Metafar.Challenge.Infrastructure.Exceptions;
-using Metafar.Challenge.Infrastructure.Utility;
 using Metafar.Challenge.Model;
-using Metafar.Challenge.Repository.Commands;
 using Metafar.Challenge.Repository.Commands.Card;
-using Metafar.Challenge.Repository.Queries;
 using Metafar.Challenge.Repository.Queries.Card;
 using Metafar.Challenge.UseCase.Security.Queries.SignInUserByCard;
 using Microsoft.Extensions.Logging;
 using Moq;
 
-
-namespace Metafar.Challenge.UnitTest;
+namespace Metafar.Challenge.UnitTest.UseCases;
 
 public class SignInUserByCardHandlerTests
 {
