+++ Metafar.Challenge/Metafar.Challenge.WebApi/Properties/launchSettings.json
@@ -17,7 +17,7 @@
       "applicationUrl": "http://localhost:5000",
       "environmentVariables": {
         "ASPNETCORE_ENVIRONMENT": "Development",
-        "DB_CONNECTION": "Server=localhost,1433;Initial Catalog=metafar.challenge.db;Persist Security Info=False;User ID=sa;Password=Password12345;MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=True;",
+        "DB_CONNECTION": "Server=,1433;Initial Catalog=metafar.challenge.db;Persist Security Info=False;User ID=sa;Password=Password12345;MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=True;",
         "JWT_SECRET": "1a382c6e-fa61-4fa1-87ac-0f3ed5aca962",
         "JWT_ISSUER": "METAFAR",
         "JWT_AUDIENCE": "METAFAR_CHALLENGE",
