+++ Metafar.Challenge/docker-compose.yml
@@ -11,6 +11,8 @@ services:
   sqlserver:
     image: mcr.microsoft.com/mssql/server:2022-latest
     container_name: sqlserver-db
+    networks:
+      - app-metafar-network-public
     ports:
       - "1433:1433"
     environment:
@@ -23,10 +25,19 @@ services:
   metafar.challenge.webapi:
     image: metafar.challenge.webapi:v1
     container_name: metafar.challenge.webapi
+    networks:
+      - app-metafar-network-public
     depends_on:
       - sqlserver
     ports:
         - '5000:8080'
+    environment:
+      ASPNETCORE_ENVIRONMENT: Development
+      DB_CONNECTION: Server=host.docker.internal,1433;Initial Catalog=metafar.challenge.db;Persist Security Info=False;User ID=sa;Password=Password12345;MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=True;
+      JWT_SECRET: 1a382c6e-fa61-4fa1-87ac-0f3ed5aca962
+      JWT_ISSUER: METAFAR
+      JWT_AUDIENCE: METAFAR_CHALLENGE
+      JWT_EXPIRATION_MINUTES: 60
     build:
       context: .
       dockerfile: Dockerfile
